requirements.txt
telethon
cryptg
pillow
aiohttp
hachoir
pandas
pyyaml
telethon